+++ src/Driver/Display/DisplayCapability.ts
@@ -1,6 +1,7 @@
 enum DisplayCapability {
 	BRIGHTNESS,
 	SCHEDULE,
+	CPU_FAN,
 }
 
 export default DisplayCapability;