+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support proximity sensor on NEC displays (there it's called "human sensor")
 - Support reset NEC display settings as means of recovery from breaking external changes
 - Internal client-server communication re-implemented using websocket instead of HTTP requests
+- Manage compute module fan based on cpu temperature in NEC displays
 
 ### Fixed
 - Casually not loading of applet after restart app